add helper folder. split folder functions into folder.js and urls functions into urls.js
