+++ app/index.html
@@ -33,6 +33,8 @@
         </div>
       </article>
     </section>
+    <script type="text/javascript" src="./helper/folders.js"></script>
+    <script type="text/javascript" src="./helper/urls.js"></script>
     <script type="text/javascript" src="main.js"></script>
   </body>
 </html>
